pyproject.toml
[project]
name = "lingchu-bot"
version = "0.0.0-dev0"
description = "由NoneBot2驱动，符合Onebot11规范的管理机器人"
keywords = ["onebot11", "nonebot2"]
readme = "README.md"
license = "LGPL-3.0 AND GFDL-1.3"
license-files = ["LICENSE-code", "LICENSE-docs", "Repository-Policy.md"]
requires-python = ">=3.13.9, <4.0.0"
authors = [
  {name = "xinvxueyuan", email = "xinvxueyuan@yeah.net"}
]
maintainers = [
  {name = "xinvxueyuan", email = "xinvxueyuan@yeah.net"}
]
dependencies = [
    "nonebot2[fastapi]>=2.4.4",
    "nonebot2[httpx]>=2.4.4",
    "nonebot2[websockets]>=2.4.4",
    "nonebot-adapter-console>=0.9.0",
    # "nonebot2[aiohttp]>=2.4.4",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-localstore>=0.7.4",
    "nonebot-plugin-orm[aiosqlite]>=0.8.3",
    "nonebot-plugin-alconna>=0.60.4",
    "psutil>=7.2.1",
    "python-jose[cryptography]>=3.5.0",
]

[project.optional-dependencies]

[dependency-groups]
devs = [
    "ruff>=0.14.10",
    "pyright>=1.1.407",
    "bump-my-version>=1.2.6",
    "poethepoet>=0.39.0",
    "nonebot-adapter-console>=0.9.0",
]
docs = [
    "zensical>=0.0.14",
]

[project.urls]
Homepage = "https://lingchu.zone.id/"
Documentation = "https://lingchu.zone.id/"
Repository = "https://github.com/xinvxueyuan/nonebot-plugin-lingchu-bot.git"
Issues = "https://github.com/xinvxueyuan/nonebot-plugin-lingchu-bot/issues"

[tool.PEP639.Standards]
license-expression = "AGPL-3.0 AND CC-BY-SA-4.0"

[tool.poe.tasks]

[tool.nonebot]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
"@local" = []

[tool.nonebot.plugins]
nonebot-plugin-apscheduler = ["nonebot_plugin_apscheduler"]
nonebot-plugin-localstore = ["nonebot_plugin_localstore"]
nonebot-plugin-orm = ["nonebot_plugin_orm"]
"@local" = []

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
# For more rules, see https://docs.astral.sh/ruff/rules/.
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "I",     # isort
    "C90",   # mccabe
    "N",     # pep8-naming
    "PL",    # pylint
    "UP",    # pyupgrade
    "YTT",   # flake8-2020
    "ANN",   # flake8-annotations
    "ASYNC", # flake8-async
    "BLE",   # flake8-blind-except
    "FBT",   # flake8-boolean-trap
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "COM",   # flake8-commas
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "T10",   # flake8-debugger
    "ICN",   # flake8-import-conventions
    "PIE",   # flake8-pie
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "Q",     # flake8-quotes
    "RSE",   # flake8-raise
    "RET",   # flake8-return
    "SIM",   # flake8-simplify
    "SLOT",  # flake8-slots
    "TID",   # flake8-tidy-imports
    "TC",    # flake8-type-checking
    "ARG",   # flake8-unused-arguments
    "PTH",   # flake8-use-pathlib
    # "ERA",   # eradicate
    "FAST", # FastAPI
    "PERF", # Perflint
    "PGH",  # pygrep-hooks
    "FURB", # refurb
    "TRY",  # tryceratops
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E402",  # module-import-not-at-top-of-file  # nonebot2 require() violates this
    "B008",  # function-call-in-default-argument  # nonebot2 Depends() without Annotated violates this
    "UP037", # quoted-annotation
    # "RUF001",  # ambiguous-unicode-character-string
    # "RUF002",  # ambiguous-unicode-character-docstring
    # "RUF003",  # ambiguous-unicode-character-comment
    # "ANN201",  # missing-return-type-undocumented-public-function
    "ANN202",  # missing-return-type-private-function
    "ANN401",  # any-type
    "COM812",  # missing-trailing-comma
    "PLC0415", # import-outside-top-level
    "TID252",  # prefer-absolute-imports
]
allowed-confusables = [
    "，",
    "。",
    "“",
    "”",
    "：",
    "；",
    "？",
    "！",
    "【",
    "】",
    "《",
    "》",
    "…",
    "—",
    "（",
    "）",
    "、",
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pyright]
# For Pylance/Pyright configurations, see https://microsoft.github.io/pyright/#/configuration.
pythonVersion = "3.13"
pythonPlatform = "All"
typeCheckingMode = "standard"


